requirements.txt
configparser==3.7.1
EasyProcess==0.2.5
python-dotenv==0.10.1
PyVirtualDisplay==0.2.1
selenium==3.141.0
urllib3==1.24.1
